chore: Trait files have been moved
