+++ src/LionDatabase/Drivers/MySQL.php
@@ -7,13 +7,6 @@ namespace Lion\Database\Drivers;
 use Closure;
 use Lion\Database\Connection;
 use Lion\Database\Driver;
-use Lion\Database\Helpers\Interfaces\ConnectionInterfaceTrait;
-use Lion\Database\Helpers\Interfaces\ExecuteInterfaceTrait;
-use Lion\Database\Helpers\Interfaces\GetAllInterfaceTrait;
-use Lion\Database\Helpers\Interfaces\GetInterfaceTrait;
-use Lion\Database\Helpers\Interfaces\QueryInterfaceTrait;
-use Lion\Database\Helpers\Interfaces\RunInterfaceTrait;
-use Lion\Database\Helpers\Interfaces\TransactionInterfaceTrait;
 use Lion\Database\Interface\DatabaseConfigInterface;
 use Lion\Database\Interface\Drivers\InsertInterface;
 use Lion\Database\Interface\QueryInterface;
@@ -21,6 +14,13 @@ use Lion\Database\Interface\ReadDatabaseDataInterface;
 use Lion\Database\Interface\RunDatabaseProcessesInterface;
 use Lion\Database\Interface\SchemaDriverInterface;
 use Lion\Database\Interface\TransactionInterface;
+use Lion\Database\Traits\ConnectionInterfaceTrait;
+use Lion\Database\Traits\ExecuteInterfaceTrait;
+use Lion\Database\Traits\GetAllInterfaceTrait;
+use Lion\Database\Traits\GetInterfaceTrait;
+use Lion\Database\Traits\QueryInterfaceTrait;
+use Lion\Database\Traits\RunInterfaceTrait;
+use Lion\Database\Traits\TransactionInterfaceTrait;
 use PDO;
 
 /**
