+++ src/LionDatabase/Traits/TransactionInterfaceTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Lion\Database\Helpers\Interfaces;
+namespace Lion\Database\Traits;
 
 /**
  * Declare the transaction method of the interface
